+++ format/json.go
@@ -1,4 +1,4 @@
-package output
+package format
 
 import (
 	"encoding/json"
