+++ format/markdown.go
@@ -1,4 +1,4 @@
-package output
+package format
 
 import (
 	"fmt"
